backend_&ML/requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
pydantic
joblib
scikit-learn
shap